+++ pymilo/transporters/general_data_structure_transporter.py
@@ -302,7 +302,11 @@ class GeneralDataStructureTransporter(AbstractTransporter):
                         [x.dtype for x in new_list])
                     if (is_homogeneous_type):
                         if all_same([len(x) for x in new_list]):
-                            return np.asarray(new_list)
+                            try:
+                                return np.asarray(new_list)
+                            except Exception as _:
+                            # when we have a list of ndarrays with different shapes.
+                                return new_list
 
             return np.asarray(new_list, dtype=object)
         else: