+++ pymilo/utils/test_pymilo.py
@@ -15,6 +15,7 @@ from ..chains.linear_model_chain import is_linear_model
 from ..chains.neural_network_chain import is_neural_network
 from ..chains.decision_tree_chain import is_decision_tree
 from ..chains.clustering_chain import is_clusterer
+from ..chains.naive_bayes_chain import is_naive_bayes
 
 from ..pymilo_param import EXPORTED_MODELS_PATH
 
@@ -36,6 +37,8 @@ def pymilo_export_path(model):
         model_type = "DECISION_TREE"
     elif is_clusterer(model):
         model_type = "CLUSTERING"
+    elif is_naive_bayes(model):
+        model_type = "NAIVE_BAYES"
     else:
         model_type = None
     return EXPORTED_MODELS_PATH[model_type]